chore(refactor): refactor main functionality
