--- src/interfaces/IDotContainer.interface.tsx
@@ -1,12 +1,0 @@
-import { Animated } from 'react-native';
-
-export interface IDotContainer {
-  navContainerMaxSize: number;
-  arrayOfSlideIndex: number[];
-  fixDotBackgroundColor: string;
-  fixDotOpacity: number;
-  dotWidth: number;
-  _moveSlideDotX: Animated.Value;
-  _moveSlideDotMarginX: Animated.Value;
-  animatedDotBackgroundColor: string;
-}
