--- src/interfaces/INextContainer.interface.tsx
@@ -1,26 +1,0 @@
-import React from 'react';
-import type { ISlide } from 'interfaces/ISlide.interface';
-
-export interface INextContainer {
-  goToNewSlide: (
-    nextActive: number,
-    slide: ISlide,
-    setSlide: (arg0: ISlide) => void,
-    numberOfSlide: number,
-    onDone: () => void,
-    navContainerMaxSize: number,
-    dotWidth: number,
-    deviceMaxWidth: number
-  ) => void;
-  slide: ISlide;
-  setSlide: (arg0: ISlide) => void;
-  numberOfSlide: number;
-  onDone: () => void;
-  navContainerMaxSize: number;
-  dotWidth: number;
-  deviceMaxWidth: number;
-  renderNextButton: (nextLabel?: string) => React.ReactNode;
-  renderDoneButton: (doneLabel?: string) => React.ReactNode;
-  isLastSlide: boolean;
-  buttonsMaxSize: number;
-}
